requirements.txt
anyio
asyncclick
bleak
boto3==1.35.38
rangehttpserver
littlefs-python==0.10.3
paho-mqtt
packaging
pillow
qoi-conv
